TESB update 1
so TESB is giving us some issues so first thing first I Installed the
latest version of TESB and formatted it to our files, split its
interaction pages as we wanted to continue to run with some on and some
off (found that to be a hilarious error), added my create-unit fix to
his TESB-create-unit (his was just a duplicate of what create-unit was
prior to my edit of it).  everything else is in just like it should
be.... could see that maybe some further "tag on" "tag off" coding could
be used.