V1 28 amos final fixes (#56)
* Fixes.

1. Gemset fixes;
2. Tagouts for Player controlled adventurer races;
3. a titanium source for dwarves - temporary till something better
happens.
4. phasing out liquid fire from fire snake venom... caused liquid fire
shirts...

* Missing line in vanilla_mw_adds

missing line in description of what is in the file.  nothing more.
doesn't change game play.

* Plant_grass fix

This allows the plant_grass maintain its definitions while still
allowing it to turn on/off standard grass.

* tool_item.txt corrections.

should correct all of them to contain all definitions and descriptions.

* More fixes to get to graphic switch

2 more files to update for graphics switch.

* The MDF Launcher Graphic switch edition

The graphic switch edition of MDF

* Reduced pets

edits to reduce pets and adventurer options.
the variation of the MDF launcher that operates these 2 options is also
included.

* more fixes

these files are not needed in the graphic folders, these match the base
raws.